+++ CleverPush/Source/CleverPushInstance.m
@@ -73,7 +73,7 @@
 
 @implementation CleverPushInstance
 
-NSString* const CLEVERPUSH_SDK_VERSION = @"1.30.3";
+NSString* const CLEVERPUSH_SDK_VERSION = @"1.30.7";
 
 static BOOL registeredWithApple = NO;
 static BOOL startFromNotification = NO;
@@ -609,15 +609,15 @@ static id isNil(id object) {
             unichar consentStatus = [vendorConsents characterAtIndex:iabtcfVendorConsentPosition - 1];
             BOOL hasConsent = (consentStatus == '1');
 
-            if (hasConsent) {
-                if (tcfMode == CPIabTcfModeTrackingWaitForConsent) {
-                    [self setTrackingConsent:YES];
-                }
+            if (tcfMode == CPIabTcfModeTrackingWaitForConsent) {
+                [self setTrackingConsent:hasConsent];
+            }
 
-                if (tcfMode == CPIabTcfModeSubscribeWaitForConsent) {
-                    [self setSubscribeConsent:YES];
-                }
-            } else {
+            if (tcfMode == CPIabTcfModeSubscribeWaitForConsent) {
+                [self setSubscribeConsent:hasConsent];
+            }
+
+            if (!hasConsent) {
                 [CPLog debug:@"The vendor does not have consent."];
             }
         } else {
@@ -864,15 +864,20 @@ static id isNil(id object) {
         return subscriptionId;
     }
 
-    dispatch_semaphore_t sema = dispatch_semaphore_create(0);
+    __block NSString *result = nil;
 
-    handleSubscribedInternal = ^(NSString*subscriptionIdNew) {
-        dispatch_semaphore_signal(sema);
-    };
+    dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0), ^{
+        dispatch_semaphore_t sema = dispatch_semaphore_create(0);
 
-    dispatch_semaphore_wait(sema, DISPATCH_TIME_FOREVER);
+        handleSubscribedInternal = ^(NSString *subscriptionIdNew) {
+            result = subscriptionIdNew;
+            dispatch_semaphore_signal(sema);
+        };
 
-    return subscriptionId;
+        dispatch_semaphore_wait(sema, DISPATCH_TIME_FOREVER);
+    });
+
+    return result;
 }
 
 #pragma mark - perform on main thread.
@@ -1055,6 +1060,8 @@ static id isNil(id object) {
 }
 
 - (void)subscribe:(CPHandleSubscribedBlock _Nullable)subscribedBlock failure:(CPFailureBlock _Nullable)failureBlock skipTopicsDialog:(BOOL)skipTopicsDialog {
+    __block BOOL successBlockCalled = NO;
+
     void(^handleSubscribe)(void) = ^{
         hasCalledSubscribe = YES;
         if (@available(iOS 10.0,*)) {
@@ -1096,14 +1103,24 @@ static id isNil(id object) {
                                 if (subscribedBlock) {
                                     [self getSubscriptionId:^(NSString* subscriptionId) {
                                         if (subscriptionId != nil && ![subscriptionId isKindOfClass:[NSNull class]] && ![subscriptionId isEqualToString:@""]) {
-                                            subscribedBlock(subscriptionId);
+                                            if (!successBlockCalled) {
+                                                subscribedBlock(subscriptionId);
+                                                successBlockCalled = YES;
+                                            } else {
+                                                [CPLog debug:@"CleverPushInstance: subscribe: Subscription callback already invoked."];
+                                            }
                                         } else {
                                             [CPLog debug:@"CleverPushInstance: subscribe: There is no subscription for CleverPush SDK."];
                                         }
                                     }];
                                 }
                             } else if (subscribedBlock) {
-                                subscribedBlock(subscriptionId);
+                                if (!successBlockCalled) {
+                                    subscribedBlock(subscriptionId);
+                                    successBlockCalled = YES;
+                                } else {
+                                    [CPLog debug:@"CleverPushInstance: subscribe: Subscription callback already invoked."];
+                                }
                             }
                         } else if (failureBlock) {
                             failureBlock([NSError errorWithDomain:@"com.cleverpush" code:410 userInfo:@{NSLocalizedDescriptionKey:@"Can not subscribe because notifications have been disabled by the user. You can call CleverPush.setIgnoreDisabledNotificationPermission(true) to still allow subscriptions, e.g. for silent pushes."}]);
@@ -1529,11 +1546,14 @@ static id isNil(id object) {
                     listener(subscriptionId);
                 }
                 pendingSubscriptionListeners = [NSMutableArray new];
-            }
 
-            if (successBlock) {
-                dispatch_async(dispatch_get_main_queue(), ^{
-                    successBlock();
+                static dispatch_once_t onceToken;
+                dispatch_once(&onceToken, ^{
+                    if (successBlock) {
+                        dispatch_async(dispatch_get_main_queue(), ^{
+                            successBlock();
+                        });
+                    }
                 });
             }
         } onFailure:^(NSError* error) {
@@ -1624,6 +1644,8 @@ static id isNil(id object) {
     if (!SYSTEM_VERSION_GREATER_THAN_OR_EQUAL_TO(@"10.0")) {
         [CleverPush handleNotificationReceived:messageDict isActive:NO];
     }
+
+    [self handleSilentNotificationReceivedWithAppBanner:messageDict];
 
     return startedBackgroundJob;
 }
@@ -1753,6 +1775,27 @@ static id isNil(id object) {
     handleNotificationOpened(result);
 }
 
+- (void)handleSilentNotificationReceivedWithAppBanner:(NSDictionary* _Nullable)messageDict {
+    NSDictionary* notification = [messageDict cleverPushDictionaryForKey:@"notification"];
+
+    if (!notification) {
+        return;
+    }
+
+    NSString* notificationId = [notification cleverPushStringForKey:@"_id"];
+
+    if ([CPUtils isEmpty:notificationId]) {
+        return;
+    }
+
+    NSString* appBanner = [notification cleverPushStringForKey:@"appBanner"];
+    bool isSilent = [notification objectForKey:@"silent"] != nil && ![[notification objectForKey:@"silent"] isKindOfClass:[NSNull class]] && [[notification objectForKey:@"silent"] boolValue];
+
+    if (![CPUtils isNullOrEmpty:appBanner] && isSilent) {
+        [CPAppBannerModuleInstance setSilentPushAppBannersIds:appBanner notificationId:notificationId];
+    }
+}
+
 #pragma mark - Handle notification actions buttons events
 - (NSMutableDictionary* _Nullable)handleActionInNotification:(NSDictionary* _Nullable)notificationPayload
                                         withAction:(NSString* _Nullable)actionIdentifier
@@ -1804,7 +1847,11 @@ static id isNil(id object) {
         } else {
             [UNUserNotificationCenter.currentNotificationCenter getDeliveredNotificationsWithCompletionHandler:^(NSArray<UNNotification*>*notifications) {
                 dispatch_async(dispatch_get_main_queue(), ^{
-                    [[UIApplication sharedApplication] setApplicationIconBadgeNumber:[notifications count]];
+                    if (@available(iOS 16.0, *)) {
+                        [[UNUserNotificationCenter currentNotificationCenter] setBadgeCount:[notifications count] withCompletionHandler:nil];
+                    } else {
+                        [[UIApplication sharedApplication] setApplicationIconBadgeNumber:[notifications count]];
+                    }
                 });
             }];
         }
@@ -2931,25 +2978,27 @@ static id isNil(id object) {
             NSDictionary*event = [channelEvents objectAtIndex:eventIndex];
             NSString*eventId = [event cleverPushStringForKey:@"_id"];
 
-            [self getSubscriptionId:^(NSString* subscriptionId) {
-                if (subscriptionId == nil) {
-                    [CPLog debug:@"CleverPushInstance: trackEvent: There is no subscription for CleverPush SDK."];
-                }
-                NSMutableURLRequest* request = [[CleverPushHTTPClient sharedClient] requestWithMethod:HTTP_POST path:@"subscription/conversion"];
-                NSDictionary* dataDic = [NSDictionary dictionaryWithObjectsAndKeys:
-                                         channelId, @"channelId",
-                                         eventId, @"eventId",
-                                         subscriptionId, @"subscriptionId",
-                                         isNil(properties), @"properties",
-                                         isNil(lastClickedSessionNotificationId), @"notificationId",
-                                         nil];
-
-                NSData* postData = [NSJSONSerialization dataWithJSONObject:dataDic options:0 error:nil];
-                [request setHTTPBody:postData];
-
-                [self enqueueRequest:request onSuccess:^(NSDictionary* results) {
-
-                } onFailure:nil];
+            [self waitForTrackingConsent:^{
+                [self getSubscriptionId:^(NSString* subscriptionId) {
+                    if (subscriptionId == nil) {
+                        [CPLog debug:@"CleverPushInstance: trackEvent: There is no subscription for CleverPush SDK."];
+                    }
+                    NSMutableURLRequest* request = [[CleverPushHTTPClient sharedClient] requestWithMethod:HTTP_POST path:@"subscription/conversion"];
+                    NSDictionary* dataDic = [NSDictionary dictionaryWithObjectsAndKeys:
+                                             channelId, @"channelId",
+                                             eventId, @"eventId",
+                                             subscriptionId, @"subscriptionId",
+                                             isNil(properties), @"properties",
+                                             isNil(lastClickedSessionNotificationId), @"notificationId",
+                                             nil];
+                    
+                    NSData* postData = [NSJSONSerialization dataWithJSONObject:dataDic options:0 error:nil];
+                    [request setHTTPBody:postData];
+                    
+                    [self enqueueRequest:request onSuccess:^(NSDictionary* results) {
+                        
+                    } onFailure:nil];
+                }];
             }];
 
             [CPAppBannerModule setCurrentEventId:eventId];
@@ -3559,6 +3608,14 @@ static id isNil(id object) {
     localEventTrackingRetentionDays = days;
 }
 
+- (void)setBadgeCount:(NSInteger)count {
+    if (@available(iOS 16.0, *)) {
+        [[UNUserNotificationCenter currentNotificationCenter] setBadgeCount:count withCompletionHandler:nil];
+    } else {
+        [UIApplication sharedApplication].applicationIconBadgeNumber = count;
+    }
+}
+
 - (NSString* _Nullable)getApiEndpoint {
     return apiEndpoint;
 }
@@ -3577,6 +3634,15 @@ static id isNil(id object) {
 
 - (int)getLocalEventTrackingRetentionDays {
     return localEventTrackingRetentionDays;
+}
+
+- (void)getBadgeCount:(void (^ _Nullable)(NSInteger))completionHandler {
+    [[UNUserNotificationCenter currentNotificationCenter] getDeliveredNotificationsWithCompletionHandler:^(NSArray<UNNotification *> * _Nonnull notifications) {
+        NSInteger badgeCount = [notifications count];
+        dispatch_async(dispatch_get_main_queue(), ^{
+            completionHandler(badgeCount);
+        });
+    }];
 }
 
 #pragma mark - App Banner methods