+++ CleverPush/Source/UNUserNotificationCenter+CleverPush.m
@@ -1,7 +1,6 @@
 #import <Foundation/Foundation.h>
 #import <UIKit/UIKit.h>
 #import <UserNotifications/UserNotifications.h>
-
 #import "UNUserNotificationCenter+CleverPush.h"
 #import "CleverPushSelectorHelpers.h"
 #import "CleverPush.h"